+++ btc-backend-commons/src/main/java/pl/btcgrouppl/btc/backend/commons/Constants.java
@@ -1,5 +1,7 @@
 package pl.btcgrouppl.btc.backend.commons;
 
+import org.springframework.beans.factory.annotation.Qualifier;
+
 /**
  * Created by Sebastian Mekal <sebitg@gmail.com> on 05.06.15.
  * <p>
@@ -25,5 +27,12 @@ public interface Constants {
         public static final String PROFILE_INTEGRATION_MONGO = "integration-mongo";
 
         public static final String PROFILE_DB_MONGO = "db-mongo";
+    }
+
+    public interface JSON {
+        public static final String SERIALIZER_CLEAR = "serializerClear";
+        public static final String SERIALIZERS_SET = "serializersSet";
+        public static final String INTEGRATION_SERIALIZER = "integrationSerializer";
+        public static final String INTEGRATION_DESERIALIZER = "integrationDeserializer";
     }
 }
