+++ btc-backend-commons/src/main/java/pl/btcgrouppl/btc/backend/commons/cqrs/models/exceptions/CqrsException.java
