+++ btc-backend-commons/src/main/java/pl/btcgrouppl/btc/backend/commons/integration/models/IntegrationMessageAware.java
@@ -1,6 +1,6 @@
-package pl.btcgrouppl.btc.backend.commons.integration;
+package pl.btcgrouppl.btc.backend.commons.integration.models;
 
-import pl.btcgrouppl.btc.backend.commons.integration.models.IntegrationMessage;
+import pl.btcgrouppl.btc.backend.commons.integration.models.pojos.IntegrationMessage;
 
 /**
  * Created by Sebastian Mekal <sebitg@gmail.com> on 01.06.15.
