+++ btc-backend-commons/src/test/java/pl/btcgrouppl/btc/backend/commons/test/util/cqrs/TestAsyncCommand.java
@@ -1,6 +1,6 @@
 package pl.btcgrouppl.btc.backend.commons.test.util.cqrs;
 
-import pl.btcgrouppl.btc.backend.commons.cqrs.annotations.CommandAnnotation;
+import pl.btcgrouppl.btc.backend.commons.cqrs.models.annotations.CommandAnnotation;
 
 /**
  * Created by Sebastian Mekal <sebitg@gmail.com> on 24.05.15.
