+++ btc-backend-commons/src/test/java/pl/btcgrouppl/btc/backend/commons/test/util/cqrs/TestCommand2.java
@@ -1,7 +1,7 @@
 package pl.btcgrouppl.btc.backend.commons.test.util.cqrs;
 
 import lombok.Data;
-import pl.btcgrouppl.btc.backend.commons.cqrs.annotations.CommandAnnotation;
+import pl.btcgrouppl.btc.backend.commons.cqrs.models.annotations.CommandAnnotation;
 
 /**
  * Created by Sebastian Mekal <sebitg@gmail.com> on 21.05.15.
